pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sqllog"
version = "1.0.0"
description = "High-reliability PLC to SQL Server data logger"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "SQLlog Contributors"}
]
keywords = ["plc", "sql", "logging", "allen-bradley", "industrial", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Logging",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
]

dependencies = [
    "pycomm3>=1.2.0",
    "pyodbc>=5.0.0",
    "loguru>=0.7.0",
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "pywin32>=306",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
build = [
    "pyinstaller>=6.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/SQLlog"
Documentation = "https://github.com/yourusername/SQLlog#readme"
Repository = "https://github.com/yourusername/SQLlog.git"
Issues = "https://github.com/yourusername/SQLlog/issues"

[project.scripts]
sqllog = "src.main:main"
sqllog-test = "test_connections:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | venv
    | build
    | dist
)/
'''

[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", "build", "dist", "venv", ".venv"]
ignore = ["E203", "W503"]
